+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PollingViewModel.swift
@@ -15,14 +15,14 @@ import UIKit
 class PollingViewModel {
 
     let paymentMethodType: STPPaymentMethodType
-    let supportedPaymentMethods: [STPPaymentMethodType] = [.UPI, .blik, .paynow]
+    let supportedPaymentMethods: [STPPaymentMethodType] = [.UPI, .blik, .paynow, .promptPay]
     lazy var CTA: String = {
         switch paymentMethodType {
         case .UPI:
             return .Localized.open_upi_app
         case .blik:
             return .Localized.blik_confirm_payment
-        case .paynow:
+        case .paynow, .promptPay:
             return .Localized.paynow_confirm_payment
         default:
             fatalError("Polling CTA has not been implemented for \(paymentMethodType)")
@@ -34,7 +34,7 @@ class PollingViewModel {
             return Date().addingTimeInterval(60 * 5) // 5 minutes
         case .blik:
             return Date().addingTimeInterval(60) // 60 seconds
-        case .paynow:
+        case .paynow, .promptPay:
             return Date().addingTimeInterval(60 * 60) // 1 hour
         default:
             fatalError("Polling deadline has not been implemented for \(paymentMethodType)")
@@ -42,7 +42,7 @@ class PollingViewModel {
     }()
     var retryInterval: TimeInterval {
         switch paymentMethodType {
-        case .blik, .paynow:
+        case .blik, .paynow, .promptPay:
             return 1
         case .UPI:
             return 10